requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
python-dateutil==2.8.2
pytz==2024.1 